COMP: fixing compile errors and working for Windows (Tested on VC++ 2010 / Windows 7)

+++ Testing/igtlutil/igtl_ndarray_test.c
@@ -76,7 +76,7 @@ int main( int argc, char * argv [] )
     }
 
   /** Allocate memory for pack **/
-  body_size = igtl_ndarray_get_size(&info, IGTL_TYPE_PREFIX_NONE);
+  body_size = (size_t)igtl_ndarray_get_size(&info, IGTL_TYPE_PREFIX_NONE);
   body = malloc(body_size);
 
   if (body == NULL)
